pillow-imageio-compressor/requirements.txt
Pillow==11.3.0
imageio==2.37.0
